cpoc-web/requirements.txt
Django==2.0
djangorestframework==3.8
django-filter==1.0.4
Pillow==4.2.0
requests==2.18.1
pytz==2017.2
psycopg2==2.7.1
git+https://github.com/prontotools/Official-API-Python-Library.git
Logentries==0.17
raven==6.1.0
Celery==4.1.1
redis==2.10.5
flower==0.9.2
django-haystack==2.6.1
elasticsearch>=2.0.0,<3.0.0
drf-haystack==1.6.1
arrow==0.10.0
django_celery_beat==1.1.1
python-intercom==3.1.0
